3rd commit
The biggest change for a while, I returned to my page after some time and remade most of the stuff. A lot of code is borrowed but I made sure not to steal any. I was thinking about implementing WordPress for blog pages but that was too complicated for me so I learned CSS grid instead. A 4th commit is coming soon, expect me doing some little touches and maybe making the code more slick.